Instructions for how to improve this checklist.
I'm trying to avoid having kibitzers send me email saying, "... why don't you ..."

Because it's on GitHub, anyone with an idea about how to make this better can fork it and fix it to be the way he wants. I can then accept or reject the pull request containing the fix.